Bug fix in TSStepNumbes and initial for TSSimulator
A few fixes for the Set-TSStepNumbers script.  Broke running remove of Primary site server.

Initial build of the task sequence simulator Invoke-TSSimulator.ps1 to attempt to simulate the execution of a task sequence.  This release works with handling the continue on error functionality when steps give unknown exit codes.  Next steps will be to incorporate basic variable handling.